Move the updateMovieFile method call
Having this call inside the for loop would make the algorithm take way
longer than if it's at the end of the the execution.

I ran a couple tests to see what was the difference; converting 947 png
formatted frames of 960 by 540 would take:

    * updateMovieFile outside the for loop
        real    0m46.974s
        user    0m11.370s
        sys 0m5.765s

    * updateMovieFile inside the for loop
        real    4m47.004s <--- This is crazy
        user    0m11.265s
        sys 0m5.765s

The timing in this benchmarks were done using the `time` command.